requirements.txt
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.7.1
cairocffi==1.0.2
CairoSVG==2.3.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
django>=2.2.11
django-anysign==1.4
django-crispy-forms==1.8.1
django-filter==21.1
django-heroku==0.3.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
lxml==4.3.3
mccabe==0.6.1
psycopg2-binary==2.8.1
pycodestyle==2.5.0
pycparser==2.19
pydocusign==2.2
pylint==2.3.1
Pyphen==0.9.5
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.1
reportlab==3.5.21
requests==2.21.0
six==1.16.0
soupsieve==1.9.1
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.4.2
urllib3==1.24.3
WeasyPrint==47
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.11.1